s390: ECX key generation fixes.
